Version 0.0.2.2:
- Bugfix: kein Absturz bei Shell-Eingaben mehr

git-svn-id: svn://svn.code.sf.net/p/prettyos/code/trunk/Source@833 31957799-4632-4615-9793-72db6d48bcac

+++ user/shell/shell.c
@@ -11,7 +11,7 @@ void eraseFirst(char* string)
 }
 void insert(char* string, char c)
 {
-    for (size_t i = strlen(string)-1; i >= 0; i--)
+    for (int i = strlen(string)-1; i >= 0; i--)
     {
         string[i+1] = string[i];
     }
